cultural-analytics/requirements.txt
qdrant-client>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
scipy>=1.7.0